%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RCmdCheckTools-package}
\alias{RCmdCheckTools-package}
\alias{RCmdCheckTools}
\docType{package}

\title{Package RCmdCheckTools}


\description{
  Methods for running and troubleshooting R package checks
}

\author{
  Henrik Bengtsson
}

\section{License}{
  The releases of this package is licensed under
  GPL version 2 or newer.
}
\keyword{package}
